Address review: co-locate AWS tests, move byte-limit to constants.ts
- Move the AWS unit tests from lib/stdlib/aws/__tests__/ to co-located
  lib/stdlib/aws/*.test.ts (and base64/objectBytes tests to co-located), matching
  the repo's dominant convention (lib/stdlib/layout, providers, etc.).
- AWS_OBJECT_BYTE_LIMIT is a fixed safety cap, not user-tunable, so it moves to
  lib/constants.ts; objectBytes.ts and http.ts import it from there.
