+++ packages/agency-lang/lib/stdlib/aws/endpoints.test.ts
@@ -2,7 +2,7 @@ import {
   resolveAwsPartition,
   validateBucket,
   BUCKET_NAME_MAX_LENGTH,
-} from "../endpoints.js";
+} from "./endpoints.js";
 
 describe("resolveAwsPartition", () => {
   it("maps commercial, GovCloud, China, and ISO regions to exact suffixes", () => {
