+++ packages/agency-lang/lib/stdlib/objectBytes.ts
@@ -1,13 +1,7 @@
 import { failure, type ResultFailure } from "../runtime/result.js";
+import { AWS_OBJECT_BYTE_LIMIT } from "../constants.js";
 
-/**
- * One shared object-size policy for AWS object bodies. Raw downloads and decoded
- * uploads intentionally share this limit — it bounds the bytes held in memory,
- * so both directions are capped by the same number rather than two drifting ones.
- */
-export const AWS_OBJECT_BYTE_LIMIT = 10 * 1024 * 1024;
-
-/** A `failure` when `bytes` exceeds the object limit, or `null` when it fits. */
+/** A `failure` when `bytes` exceeds the AWS object limit, or `null` when it fits. */
 export function objectSizeFailure(bytes: Uint8Array): ResultFailure | null {
   if (bytes.byteLength > AWS_OBJECT_BYTE_LIMIT) {
     return failure({
